housekeep code debug

+++ data-processing/housekeep_min_to_hour.py
@@ -77,7 +77,7 @@ def compress_time(df, t_window, start_tick, tstep = 60, from_csv = True ):
             )
     if t_window:
         df = select_time_window(df, start_tick=start_tick, t_window=t_window)
-    df = df.withColumn("event_time", ((df.event_time - t0) / tstep).cast('integer') * tstep + t0)
+    df = df.withColumn("event_time", ((df.event_time.cast("long") - t0) / tstep).cast('long') * tstep + t0)
     df = df.withColumn("event_time", F.from_utc_timestamp(F.to_timestamp(df.event_time), 'UTC'))
     # t_max = df.agg({"event_time": "max"}).collect()[0][0]
